Imported 0.15.0.31
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@17229 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/msdos/kbd.c
@@ -203,7 +203,7 @@ static void queue_command_end(void) { }
 /* CPU trap to enter the main menu.  */
 static void menu_trap(ADDRESS addr, void *data)
 {
-    ui_main((char) data);
+    ui_main((char) (unsigned long) data);
 }
 
 /* Dispatch all the pending keyboard commands.  */
